pyproject.toml
[project]
name = "app-backlog-games-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Pedro Pereira", email = "pedrohspereira2000@gmail.com"},
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pymongo>=4.6.1",
    "pydantic>=2.5.3",
    "python-decouple>=3.8",
    "auth>=0.5.3",
    "python-jose>=3.3.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
